+++ crates/matrix-sdk-base/src/response_processors/mod.rs
@@ -13,42 +13,38 @@
 // limitations under the License.
 
 pub mod account_data;
+pub mod changes;
 #[cfg(feature = "e2e-encryption")]
 pub mod e2ee;
+pub mod ephemeral_events;
 #[cfg(feature = "e2e-encryption")]
 pub mod latest_event;
+pub mod notification;
+pub mod profiles;
+pub mod room;
+pub mod state_events;
+pub mod timeline;
 #[cfg(feature = "e2e-encryption")]
 pub mod verification;
 
 use std::collections::BTreeMap;
 
-#[cfg(feature = "e2e-encryption")]
-mod with_e2ee {
-    pub use super::{
-        e2ee::{e2ee_from_msc4186, e2ee_from_sync_v3},
-        latest_event::decrypt_latest_events,
-        verification::verification,
-    };
-}
 use ruma::OwnedRoomId;
-#[cfg(feature = "e2e-encryption")]
-pub use with_e2ee::*;
 
 use crate::{RoomInfoNotableUpdateReasons, StateChanges};
 
 type RoomInfoNotableUpdates = BTreeMap<OwnedRoomId, RoomInfoNotableUpdateReasons>;
 
+#[cfg_attr(test, derive(Clone))]
+#[derive(Default)]
 pub(crate) struct Context {
     pub(super) state_changes: StateChanges,
     pub(super) room_info_notable_updates: RoomInfoNotableUpdates,
 }
 
 impl Context {
-    pub fn new(
-        state_changes: StateChanges,
-        room_info_notable_updates: RoomInfoNotableUpdates,
-    ) -> Self {
-        Self { state_changes, room_info_notable_updates }
+    pub fn new(state_changes: StateChanges) -> Self {
+        Self { state_changes, room_info_notable_updates: Default::default() }
     }
 
     pub fn into_parts(self) -> (StateChanges, RoomInfoNotableUpdates) {