+++ crates/matrix-sdk/src/client/mod.rs
@@ -2213,7 +2213,7 @@ impl Client {
     /// client
     ///     .sync_with_callback(sync_settings, |response| async move {
     ///         let channel = sync_channel;
-    ///         for (room_id, room) in response.rooms.join {
+    ///         for (room_id, room) in response.rooms.joined {
     ///             for event in room.timeline.events {
     ///                 channel.send(event).await.unwrap();
     ///             }
@@ -2293,7 +2293,7 @@ impl Client {
     ///     .sync_with_result_callback(sync_settings, |response| async move {
     ///         let channel = sync_channel;
     ///         let sync_response = response?;
-    ///         for (room_id, room) in sync_response.rooms.join {
+    ///         for (room_id, room) in sync_response.rooms.joined {
     ///              for event in room.timeline.events {
     ///                  channel.send(event).await.unwrap();
     ///               }
@@ -2366,7 +2366,7 @@ impl Client {
     ///     Box::pin(client.sync_stream(SyncSettings::default()).await);
     ///
     /// while let Some(Ok(response)) = sync_stream.next().await {
-    ///     for room in response.rooms.join.values() {
+    ///     for room in response.rooms.joined.values() {
     ///         for e in &room.timeline.events {
     ///             if let Ok(event) = e.raw().deserialize() {
     ///                 println!("Received event {:?}", event);
@@ -2749,12 +2749,12 @@ pub(crate) mod tests {
         let retry_timeout = Duration::from_secs(5);
         let server = MockServer::start().await;
         let client = test_client_builder(Some(server.uri()))
-            .request_config(RequestConfig::new().retry_timeout(retry_timeout))
+            .request_config(RequestConfig::new().max_retry_time(retry_timeout))
             .build()
             .await
             .unwrap();
 
-        assert!(client.request_config().retry_timeout.unwrap() == retry_timeout);
+        assert!(client.request_config().max_retry_time.unwrap() == retry_timeout);
 
         Mock::given(method("POST"))
             .and(path("/_matrix/client/r0/login"))