+++ crates/matrix-sdk/src/room/mod.rs
@@ -158,7 +158,10 @@ use crate::{
     BaseRoom, Client, Error, HttpResult, Result, RoomState, TransmissionProgress,
 };
 #[cfg(feature = "e2e-encryption")]
-use crate::{crypto::types::events::CryptoContextInfo, encryption::backups::BackupState};
+use crate::{
+    crypto::types::events::CryptoContextInfo, encryption::backups::BackupState,
+    room::shared_room_history::share_room_history,
+};
 
 pub mod edit;
 pub mod futures;
@@ -172,6 +175,9 @@ pub mod reply;
 
 /// Contains all the functionality for modifying the privacy settings in a room.
 pub mod privacy_settings;
+
+#[cfg(feature = "e2e-encryption")]
+mod shared_room_history;
 
 /// A struct containing methods that are common for Joined, Invited and Left
 /// Rooms
@@ -1800,6 +1806,21 @@ impl Room {
         Ok(())
     }
 
+    /// Share any shareable E2EE history in this room with the given recipient,
+    /// as per [MSC4268].
+    ///
+    /// [MSC4268]: https://github.com/matrix-org/matrix-spec-proposals/pull/4268
+    ///
+    /// This is temporarily exposed for integration testing as part of
+    /// experimental work on history sharing. In future, it will be combined
+    /// with sending an invite.
+    #[cfg(feature = "e2e-encryption")]
+    #[doc(hidden)]
+    #[instrument(skip_all, fields(room_id = ?self.room_id(), ?user_id))]
+    pub async fn share_history<'a>(&'a self, user_id: &UserId) -> Result<()> {
+        share_room_history(self, user_id.to_owned()).await
+    }
+
     /// Wait for the room to be fully synced.
     ///
     /// This method makes sure the room that was returned when joining a room
@@ -2991,17 +3012,20 @@ impl Room {
     /// Get the membership details for the current user.
     ///
     /// Returns:
-    ///     - If the current user was present in the room, a tuple of the
-    ///       current user's [`RoomMember`] info and the member info of the
-    ///       sender of that member event.
+    ///     - If the user was present in the room, a
+    ///       [`RoomMemberWithSenderInfo`] containing both the user info and the
+    ///       member info of the sender of the `m.room.member` event.
     ///     - If the current user is not present, an error.
-    pub async fn own_membership_details(&self) -> Result<(RoomMember, Option<RoomMember>)> {
-        let Some(own_member) = self.get_member_no_sync(self.own_user_id()).await? else {
+    pub async fn member_with_sender_info(
+        &self,
+        user_id: &UserId,
+    ) -> Result<RoomMemberWithSenderInfo> {
+        let Some(member) = self.get_member_no_sync(user_id).await? else {
             return Err(Error::InsufficientData);
         };
 
         let sender_member =
-            if let Some(member) = self.get_member_no_sync(own_member.event().sender()).await? {
+            if let Some(member) = self.get_member_no_sync(member.event().sender()).await? {
                 // If the sender room member info is already available, return it
                 Some(member)
             } else if self.are_members_synced() {
@@ -3009,12 +3033,12 @@ impl Room {
                 None
             } else if self.sync_members().await.is_ok() {
                 // Try getting the sender room member info again after syncing
-                self.get_member_no_sync(own_member.event().sender()).await?
+                self.get_member_no_sync(member.event().sender()).await?
             } else {
                 None
             };
 
-        Ok((own_member, sender_member))
+        Ok(RoomMemberWithSenderInfo { room_member: member, sender_info: sender_member })
     }
 
     /// Forget this room.
@@ -3860,9 +3884,19 @@ impl EventSource for &Room {
 #[error("out of range conversion attempted")]
 pub struct TryFromReportedContentScoreError(());
 
+/// Contains the current user's room member info and the optional room member
+/// info of the sender of the `m.room.member` event that this info represents.
+#[derive(Debug)]
+pub struct RoomMemberWithSenderInfo {
+    /// The actual room member.
+    pub room_member: RoomMember,
+    /// The info of the sender of the event `room_member` is based on, if
+    /// available.
+    pub sender_info: Option<RoomMember>,
+}
+
 #[cfg(all(test, not(target_arch = "wasm32")))]
 mod tests {
-    use assert_matches2::assert_matches;
     use matrix_sdk_base::{store::ComposerDraftType, ComposerDraft};
     use matrix_sdk_test::{
         async_test, event_factory::EventFactory, test_json, JoinedRoomBuilder, StateTestEvent,
@@ -4103,7 +4137,7 @@ mod tests {
 
         // Since there is no member event for the own user, the method fails.
         // This should never happen in an actual room.
-        let error = room.own_membership_details().await.err();
+        let error = room.member_with_sender_info(client.user_id().unwrap()).await.err();
         assert!(error.is_some());
     }
 
@@ -4120,14 +4154,16 @@ mod tests {
         let room = server.sync_room(&client, joined_room_builder).await;
 
         // When we load the membership details
-        let ret = room.own_membership_details().await;
-        assert_matches!(ret, Ok((member, sender)));
+        let ret = room
+            .member_with_sender_info(client.user_id().unwrap())
+            .await
+            .expect("Room member info should be available");
 
         // We get the member info for the current user
-        assert_eq!(member.event().user_id(), user_id);
+        assert_eq!(ret.room_member.event().user_id(), user_id);
 
         // But there is no info for the sender
-        assert!(sender.is_none());
+        assert!(ret.sender_info.is_none());
     }
 
     #[async_test]
@@ -4143,15 +4179,17 @@ mod tests {
         let room = server.sync_room(&client, joined_room_builder).await;
 
         // When we load the membership details
-        let ret = room.own_membership_details().await;
-        assert_matches!(ret, Ok((member, sender)));
+        let ret = room
+            .member_with_sender_info(client.user_id().unwrap())
+            .await
+            .expect("Room member info should be available");
 
         // We get the current user's member info
-        assert_eq!(member.event().user_id(), user_id);
+        assert_eq!(ret.room_member.event().user_id(), user_id);
 
         // And the sender has the same info, since it's also the current user
-        assert!(sender.is_some());
-        assert_eq!(sender.unwrap().event().user_id(), user_id);
+        assert!(ret.sender_info.is_some());
+        assert_eq!(ret.sender_info.unwrap().event().user_id(), user_id);
     }
 
     #[async_test]
@@ -4171,15 +4209,17 @@ mod tests {
         let room = server.sync_room(&client, joined_room_builder).await;
 
         // When we load the membership details
-        let ret = room.own_membership_details().await;
-        assert_matches!(ret, Ok((member, sender)));
+        let ret = room
+            .member_with_sender_info(client.user_id().unwrap())
+            .await
+            .expect("Room member info should be available");
 
         // We get the current user's member info
-        assert_eq!(member.event().user_id(), user_id);
+        assert_eq!(ret.room_member.event().user_id(), user_id);
 
         // And also the sender info from the events received in the sync
-        assert!(sender.is_some());
-        assert_eq!(sender.unwrap().event().user_id(), sender_id);
+        assert!(ret.sender_info.is_some());
+        assert_eq!(ret.sender_info.unwrap().event().user_id(), sender_id);
     }
 
     #[async_test]
@@ -4204,14 +4244,16 @@ mod tests {
             .await;
 
         // We get the current user's member info
-        let ret = room.own_membership_details().await;
-        assert_matches!(ret, Ok((member, sender)));
+        let ret = room
+            .member_with_sender_info(client.user_id().unwrap())
+            .await
+            .expect("Room member info should be available");
 
         // We get the current user's member info
-        assert_eq!(member.event().user_id(), user_id);
+        assert_eq!(ret.room_member.event().user_id(), user_id);
 
         // And also the sender info from the /members endpoint
-        assert!(sender.is_some());
-        assert_eq!(sender.unwrap().event().user_id(), sender_id);
+        assert!(ret.sender_info.is_some());
+        assert_eq!(ret.sender_info.unwrap().event().user_id(), sender_id);
     }
 }